Fix Draw annotation for non-default enums
DrawType.PopupList was using the enum's value as an index.
This works fine for sequential enums that start at zero, which is the default enum definition.
It doesn't work for non-sequential enums or enums that start at a number other than zero.
The incompatibility is fixed by finding the index of the currently selected value and passing that to PopupToggleGroup instead of the value itself.